+++ src/mobster/cmd/generate/oci_image/base_images_dockerfile.py
@@ -18,9 +18,10 @@ from spdx_tools.spdx.model.relationship import Relationship, RelationshipType
 from mobster.cmd.generate.oci_image.cyclonedx_wrapper import CycloneDX1BomWrapper
 from mobster.cmd.generate.oci_image.spdx_utils import find_spdx_root_packages_spdxid
 from mobster.image import Image
-from mobster.oci import make_oci_auth_file, run_async_subprocess
+from mobster.oci import make_oci_auth_file
 from mobster.sbom.cyclonedx import get_component
 from mobster.sbom.spdx import get_image_package
+from mobster.utils import run_async_subprocess
 
 LOGGER = logging.getLogger(__name__)
