+++ src/mobster/utils.py
@@ -1,6 +1,12 @@
 """A place for utility functions used across the application."""
 
+import asyncio
+import logging
+import os
+import platform
 import re
+
+LOGGER = logging.getLogger(__name__)
 
 
 def normalize_file_name(current_name: str) -> str:
@@ -14,3 +20,47 @@ def normalize_file_name(current_name: str) -> str:
         str: The normalized file name.
     """
     return re.sub(r'[<>:"/\\|?*]', "_", current_name)
+
+
+async def run_async_subprocess(
+    cmd: list[str], env: dict[str, str] | None = None, retry_times: int = 0
+) -> tuple[int, bytes, bytes]:
+    """Run command in subprocess asynchronously.
+
+    Args:
+        cmd: Command to run in subprocess.
+        env: Environment dictionary.
+        retry_times: Number of retries if the process ends with non-zero return code.
+
+    Returns:
+        tuple[int, bytes, bytes]: Return code, stdout, and stderr.
+    """
+    if retry_times < 0:
+        raise ValueError("Retry count cannot be negative.")
+
+    cmd_env = dict(os.environ)
+    if env:
+        cmd_env.update(env)
+
+    # do this to avoid unbound warnings,
+    # the loop always runs at least once, so they're always set
+    code, stdout, stderr = 0, b"", b""
+
+    for _ in range(1 + retry_times):
+        proc = await asyncio.create_subprocess_exec(
+            *cmd,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE,
+            env=cmd_env,
+        )
+
+        stdout, stderr = await proc.communicate()
+        assert (
+            proc.returncode is not None
+        )  # can't be None after proc.communicate is awaited
+        code = proc.returncode
+        if code == 0:
+            return code, stdout, stderr
+
+    return code, stdout, stderr
+
